# doc/openapi.yaml
openapi: "3.0.3"
info:
  title: AiR ORCHESTRATOR API
  description: |
    REST API сервиса AiR ORCHESTRATOR. Большинство защищённых маршрутов требуют
    заголовок `Authorization: Bearer <token>`. WebSocket-маршруты принимают
    токен через query-параметр `token` или через `Sec-WebSocket-Protocol`.
  version: "1.0.0"
servers:
  - url: https://localhost:8443
    description: Локальная разработка (TLS)
  - url: http://localhost:8080
    description: Production (за nginx)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /avito/status:
    get:
      tags: [avito]
      summary: Статус Avito-интеграции
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /avito/auth/url:
    post:
      tags: [avito]
      summary: Сгенерировать URL авторизации Avito
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /avito/disconnect:
    post:
      tags: [avito]
      summary: Отключить Avito-интеграцию
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /avito/chats:
    get:
      tags: [avito]
      summary: Получить чаты Avito
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /avito/subscriptions:
    get:
      tags: [avito]
      summary: Список webhook-подписок Avito
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /avito/subscribe:
    post:
      tags: [avito]
      summary: Подписаться на webhook Avito
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
  /avito/unsubscribe:
    post:
      tags: [avito]
      summary: Отписаться от webhook Avito
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK

  # ─────────────── WEBSOCKETS ───────────────
...